+++ modules/core/builtin/cpp/namelengthmaxBuiltin.cpp
@@ -1,0 +1,26 @@
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of the Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// SPDX-License-Identifier: LGPL-3.0-or-later
+// LICENCE_BLOCK_END
+//=============================================================================
+#include "namelengthmaxBuiltin.hpp"
+#include "Error.hpp"
+#include "Keywords.hpp"
+#include "CheckerHelpers.hpp"
+//=============================================================================
+using namespace Nelson;
+//=============================================================================
+ArrayOfVector
+Nelson::CoreGateway::namelengthmaxBuiltin(int nLhs, const ArrayOfVector& argIn)
+{
+    ArrayOfVector retval;
+    nargincheck(argIn, 0, 0);
+    nargoutcheck(nLhs, 0, 1);
+    retval << ArrayOf::doubleConstructor(IDENTIFIER_LENGTH_MAX);
+    return retval;
+}
+//=============================================================================