+++ modules/core/src/cpp/EvaluateCommand.cpp
@@ -11,11 +11,11 @@
 #include "EvaluateCommand.hpp"
 #include "Error.hpp"
 #include "i18n.hpp"
-#include "PredefinedErrorMessages.hpp"
 #include "EvaluateInterface.hpp"
 #include "ParserInterface.hpp"
 #include "characters_encoding.hpp"
 #include "NelsonPrint.hpp"
+#include "PredefinedErrorMessages.hpp"
 //=============================================================================
 namespace Nelson {
 //=============================================================================